Refactoring config
