+++ package.json
@@ -3,7 +3,6 @@
   "version": "1.0.0",
   "license": "MIT",
   "dependencies": {
-    "@types/node": "^12.7.2",
     "apollo-server-koa": "^2.9.0",
     "dotenv": "^8.1.0",
     "graphql": "^14.5.1",
@@ -20,6 +19,7 @@
     "typescript": "^3.5.3"
   },
   "devDependencies": {
+    "@types/node": "^12.7.2",
     "jshint": "^2.10.2",
     "jslint": "^0.12.1",
     "nodemon": "^1.19.1"
