+++ src/initializers/index.ts
@@ -1,0 +1,3 @@
+export * from './config';
+export * from './apollo-server';
+export * from './database';
